pyproject.toml
[project]
name = "agentic-ai-workshop"
version = "0.1.0"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-community>=0.3.0",
    "langsmith>=0.1.0",
    "anthropic>=0.40.0",
    "duckduckgo-search>=6.0.0",
    "ddgs>=6.0.0",
    "python-dotenv>=1.0.0",
]